refactor(gateway): split runner inventory runtime
